+++ routes/task.js
@@ -11,31 +11,31 @@ const Permissions = require('../auth/permissions');
  * @param req.params.parentId: the id of the parent
  * @return added task
  */
-router.post('/:parentType/:parentId', verifyToken, (req,res)=>{
+router.post('/:parentType/:parentId', verifyToken, async (req,res)=>{
     const parentId = req.params.parentId;
-    const parentType = req.params.parentType;
-    Branch.getParentByType(parentType,parentId).then((result)=>{
-        if(!Permissions.checkObjectPermissions(result.accountId,req.userId)){
+    const parentTypeString = req.params.parentType;
+    try {
+        const parentType = Branch.getParentType(parentTypeString);
+        const result = await Branch.getParentByType(parentType,parentId)
+        if (Permissions.checkObjectPermissions(result.accountId, req.userId)) {
+            const parent = result;
+            const task = req.body;
+            task.accountId = req.userId;
+            task.parent = parent._id;
+            task.parentTypeString = req.params.parentType;
+            const tsk = await Branch.createTask(task)
+            const taskList = parent.tasks ? parent.tasks : parent.subTasks;
+            taskList.push(tsk._id);
+            const parentType = Branch.getParentType(parentTypeString);
+            await Branch.updateParent(parentType, parentId, parent)
+            res.status(200).send(tsk);
+        } else {
             res.status(403).send('You do not have access to that parent object');
         }
-        const parent = result;
-        const task = req.body;
-        task.accountId=req.userId;
-        task.parent = parent._id;
-        task.parentType = req.params.parentType;
-        Branch.createTask(task,(err,tsk)=>{
-            if(err){
-                res.status(500).send("Error creating task: " + err );
-            }
-            const taskList = parent.tasks?parent.tasks:parent.subTasks;
-            taskList.push(tsk._id);
-            Branch.updateParent(parentType, parentId, parent).then((prnt)=>{
-                res.status(200).send(tsk);
-            })
-        })
-    }, (error)=>{
-        res.status(500).send(error);
-    })
+    } catch (e) {
+        res.status(500).send({'err': `there was an error: ${e.message}`})
+    }
+
 });
 
 /**
@@ -45,15 +45,17 @@ router.post('/:parentType/:parentId', verifyToken, (req,res)=>{
  */
 router.get('/:taskId', verifyToken, async (req,res)=>{
     const taskId = req.params.taskId;
-    try{
-        if(!(await Branch.verifyOwnership(Branch.Task,taskId,req.userId))) {
+    try {
+        if (await Branch.verifyOwnership(Branch.Task, taskId, req.userId)) {
+            const task = await Branch.getTaskRecursive(taskId);
+            res.status(200).send(task);
+        } else {
             res.status(403).send({"err": "You are not authorized to get that task"});
         }
-        const task = await Branch.getTaskRecursive(taskId);
-        res.status(200).send(task);
-    }  catch(err) {
-        res.status(500).send({'err':`error retrieving task: ${err.message}`});
+    } catch(err){
+        res.status(500).send({'err': `error retrieving task: ${err.message}`});
     }
+
 })
 
 /**
@@ -63,26 +65,24 @@ router.get('/:taskId', verifyToken, async (req,res)=>{
  */
 router.delete('/:id', verifyToken, async (req,res)=>{
     const taskId = req.params.id;
-    if(!(await Branch.verifyOwnership(Branch.Task,taskId,req.userId))){
-        res.status(403).send({"err":"You are not authorized to delete that task"});
-    }
-    try{
-        const deletedTask = await Branch.deleteTaskRecursive(taskId);
-        if(deletedTask.err){
-            res.status(500).send({"err":"error deleting task" + deletedTask.err});
+    try {
+        if (await Branch.verifyOwnership(Branch.Task, taskId, req.userId)) {
+            const deletedTask = await Branch.deleteTaskRecursive(taskId);
+            const parentId = deletedTask.parentId;
+            const parentType = deletedTask.parentType;
+            const updatedParent = Branch.getParentByType(parentType, parentId);
+            const eventIndex = updatedParent.tasks.indexOf(deletedTask._id);
+            if (eventIndex === -1) {
+                throw new Error("task was not included in its parent");
+            }
+            updatedParent.tasks.splice(eventIndex);
+            await Branch.updateParent(parentType, parentId, updatedParent);
+            res.status(200).send(deletedTask);
+        } else {
+            res.status(403).send({"err": "You are not authorized to delete that task"});
         }
-        const parentId = deletedTask.parentId;
-        const parentType = deletedTask.parentType;
-        const updatedParent = Branch.getParentByType(parentType,parentId);
-        const eventIndex = updatedParent.tasks.indexOf(deletedTask._id);
-        if(eventIndex === -1){
-            throw new Error("task was not included in its parent");
-        }
-        updatedParent.tasks.splice(eventIndex);
-        await Branch.updateParent(parentType,parentId,updatedParent);
-        res.status(200).send(deletedTask);
-    } catch(err) {
-        res.status(500).send({'err':`error deleting task: ${err.message}`});
+    } catch (err) {
+        res.status(500).send({'err': `error deleting task: ${err.message}`});
     }
 
 });
@@ -97,10 +97,14 @@ router.delete('/:id', verifyToken, async (req,res)=>{
 router.delete('/:id/:newParentType/:newParentId', verifyToken, async(req,res)=>{
     const id = req.params.id;
     try {
-        const newParentType = Branch.getParentType(req.params.newParentType);
-        const newParentId = req.params.newParentId;
-        const deletedTask = await Branch.deleteTaskAndRebaseChildren(id,newParentType,newParentId);
-        res.status(200).send(deletedTask);
+        if (await Branch.verifyOwnership(Branch.Task, taskId, req.userId)) {
+            const newParentType = Branch.getParentType(req.params.newParentType);
+            const newParentId = req.params.newParentId;
+            const deletedTask = await Branch.deleteTaskAndRebaseChildren(id, newParentType, newParentId);
+            res.status(200).send(deletedTask);
+        } else {
+            res.status(403).send({"err": "You are not authorized to delete that task"});
+        }
     } catch (err) {
         res.status(500).send('error deleting category' + err.message);
     }
@@ -116,14 +120,18 @@ router.delete('/:id/:newParentType/:newParentId', verifyToken, async(req,res)=>{
 router.put('/:taskId', verifyToken, async (req,res)=>{
     const taskId = req.params.taskId;
     const update = req.body;
-    if(!(await Branch.verifyOwnership(Branch.Task, taskId,req.userId))){
-        res.status(403).send({"err":"You are not authorized to modify that task"})
+    try {
+        if (!(await Branch.verifyOwnership(Branch.Task, taskId, req.userId))) {
+            res.status(403).send({"err": "You are not authorized to modify that task"})
+        }
+        const updatedTask = await Branch.updateTask(taskId, update);
+        if (updatedTask.err) {
+            res.status(500).send({"err": "error updating task" + updatedTask.err});
+        }
+        res.status(200).send(updatedTask);
+    }catch (e) {
+        res.status(500).send('error updating category' + e.message);
     }
-    const updatedTask = await Branch.updateTask(taskId,update);
-    if(updatedTask.err){
-        res.status(500).send({"err":"error updating task" + updatedTask.err});
-    }
-    res.status(200).send(updatedTask);
 })
